requirements.txt

anyio==3.5.0
asgiref==3.5.0
beautifulsoup4==4.11.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cssselect==1.1.0
fastapi==0.75.1
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
lxml==4.8.0
newspaper3k==0.2.8
nltk==3.7
Pillow==9.1.0
pydantic==1.9.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2
starlette==0.17.1
tinysegmenter==0.3
tldextract==3.2.0
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6